requirements.txt
beautifulsoup4==4.11.1
bs4==0.0.1
cachetools==5.2.0
certifi==2022.9.24
chardet==3.0.4
click==8.1.3
Flask==2.2.2
Flask-Cors==3.0.10
Flask-Login==0.6.2
Flask-SSLify==0.1.5
google-api-core==1.16.0
google-auth==1.5.0
google-auth-httplib2==0.0.3
google-cloud-core==0.28.1
google-cloud-storage==1.10.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
gunicorn==19.7.1
httplib2==0.20.4
idna==2.7
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.21.6
pandas==1.3.5
pip==22.2.2
protobuf==4.21.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.5
requests==2.19.1
rsa==4.9
setuptools==63.4.1
six==1.16.0
soupsieve==2.3.2.post1
stockx-py-sdk==0.0.2
typing_extensions==4.4.0
urllib3==1.23
Werkzeug==2.2.2
wheel==0.37.1
zipp==3.9.0
